Removed avoid school run from call back time slot
We have discussed the logic of choosing and showing call back
time slots and decided it's better to remove 'avoid school run'.
In this scenario it's more likely that a person can answer the phone
in more flexible time frame.

Removed 'Something else' from sidebar as it may confuse users.

Further refinement of wording for buttons and back links.